Finalise first attempt at course milestones. Split out all model files into separate classes.
