+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/courses/milestones/CourseMilestoneRepository.kt
@@ -4,11 +4,10 @@ import com.pearlbailey.pearlbaileyhighschool.courses.milestones.model.CourseMile
 import com.pearlbailey.pearlbaileyhighschool.courses.milestones.model.CourseMilestoneType
 import org.springframework.data.repository.CrudRepository
 
-@Suppress("FunctionName")
 interface CourseMilestoneRepository : CrudRepository<CourseMilestone, Int> {
 
-    fun findAllByCourse_Id(courseId: Int): List<CourseMilestone>?
+    fun findAllByCourseId(courseId: Int): List<CourseMilestone>?
 
-    fun findAllByCourse_IdAndType(courseId: Int, type: CourseMilestoneType): List<CourseMilestone>?
+    fun findAllByCourseIdAndType(courseId: Int, type: CourseMilestoneType): List<CourseMilestone>?
 
 }
