+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/courses/model/CourseResponseDto.kt
@@ -1,0 +1,12 @@
+package com.pearlbailey.pearlbaileyhighschool.courses.model
+
+import com.pearlbailey.pearlbaileyhighschool.courses.milestones.model.CourseMilestoneResponse
+
+data class CourseResponseDto(
+    val name: String,
+    val teacherId: Int,
+    val departmentId: Int,
+    val description: String,
+    val courseStatus: CourseStatus,
+    val courseMilestones: List<CourseMilestoneResponse>? = null
+)
