+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/courses/model/CourseStatus.kt
@@ -1,0 +1,5 @@
+package com.pearlbailey.pearlbaileyhighschool.courses.model
+
+enum class CourseStatus {
+    PLANNED, ACTIVE, ON_HOLD, DISCONTINUED
+}
