+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/courses/model/CreateCourseResponse.kt
@@ -1,0 +1,5 @@
+package com.pearlbailey.pearlbaileyhighschool.courses.model
+
+data class CreateCourseResponse(val id: Int)
+
+fun Int.toCreateCourseResponseDto() = CreateCourseResponse(this)
