+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/department/model/CreateDepartmentResponse.kt
@@ -1,0 +1,5 @@
+package com.pearlbailey.pearlbaileyhighschool.department.model
+
+data class CreateDepartmentResponse(val id: Int)
+
+fun Int.toCreateDepartmentResponseDto() = CreateDepartmentResponse(this)
