+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/department/model/Department.kt
@@ -11,9 +11,6 @@ import javax.persistence.Id
 import javax.persistence.JoinColumn
 import javax.persistence.OneToOne
 import javax.persistence.Table
-import javax.validation.constraints.NotBlank
-import javax.validation.constraints.NotNull
-import javax.validation.constraints.Positive
 
 @Entity
 @Table(name = "departments")
@@ -31,29 +28,3 @@ class Department {
     @JoinColumn(name = "head_of_department_id", referencedColumnName = "id")
     var headOfDepartment: Teacher? = null
 }
-
-data class DepartmentResponseDto(
-    val name: String,
-    val headOfDepartment: Int
-)
-
-fun Department.toDepartmentResponseDto() = DepartmentResponseDto(name!!, this.headOfDepartment!!.id!!)
-
-data class CreateDepartmentDto(
-    @field:NotBlank val name: String,
-    @field:NotNull @field:Positive val headOfDepartmentId: Int
-)
-
-data class CreateDepartmentResponse(val id: Int)
-
-fun Int.toCreateDepartmentResponseDto() = CreateDepartmentResponse(this)
-
-fun CreateDepartmentDto.toDepartment(headOfDepartment: Teacher) = Department().apply {
-    this.name = this@toDepartment.name;
-    this.headOfDepartment = headOfDepartment
-}
-
-data class PatchDepartmentDto(
-    val name: String?,
-    val headOfDepartmentId: Int?
-)
