+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/department/model/DepartmentResponseDto.kt
@@ -1,0 +1,8 @@
+package com.pearlbailey.pearlbaileyhighschool.department.model
+
+data class DepartmentResponseDto(
+    val name: String,
+    val headOfDepartment: Int
+)
+
+fun Department.toDepartmentResponseDto() = DepartmentResponseDto(name!!, this.headOfDepartment!!.id!!)
