+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/department/model/PatchDepartmentDto.kt
@@ -1,0 +1,6 @@
+package com.pearlbailey.pearlbaileyhighschool.department.model
+
+data class PatchDepartmentDto(
+    val name: String?,
+    val headOfDepartmentId: Int?
+)
