+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/student/model/CreateStudentResponse.kt
@@ -1,0 +1,5 @@
+package com.pearlbailey.pearlbaileyhighschool.student.model
+
+data class CreateStudentResponse(val id: Int)
+
+fun Int.toCreateStudentResponseDto() = CreateStudentResponse(this)
