+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/student/model/PatchStudentDto.kt
@@ -1,0 +1,11 @@
+package com.pearlbailey.pearlbaileyhighschool.student.model
+
+import java.time.LocalDate
+
+data class PatchStudentDto(
+    val firstName: String?,
+    val middleName: String?,
+    val lastName: String?,
+    val dateOfBirth: LocalDate?,
+    val status: StudentStatus?
+)
