+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/student/model/StudentResponseDto.kt
@@ -1,0 +1,13 @@
+package com.pearlbailey.pearlbaileyhighschool.student.model
+
+import java.time.LocalDate
+
+data class StudentResponseDto(
+    val firstName: String,
+    val middleName: String?,
+    val lastName: String,
+    val dateOfBirth: LocalDate,
+    val status: StudentStatus
+)
+
+fun Student.toStudentResponseDto() = StudentResponseDto(firstName!!, middleName, lastName!!, dateOfBirth!!, status!!)
