+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/student/model/StudentStatus.kt
@@ -1,0 +1,5 @@
+package com.pearlbailey.pearlbaileyhighschool.student.model
+
+enum class StudentStatus {
+    PROSPECT, ENROLLED, GRADUATED, SUSPENDED, EXPELLED, DECEASED
+}
