+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/teacher/model/CreateTeacherResponse.kt
@@ -1,0 +1,5 @@
+package com.pearlbailey.pearlbaileyhighschool.teacher.model
+
+data class CreateTeacherResponse(val id: Int)
+
+fun Int.toCreateTeacherResponseDto() = CreateTeacherResponse(this)
