+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/teacher/model/Teacher.kt
@@ -7,9 +7,6 @@ import javax.persistence.GeneratedValue
 import javax.persistence.GenerationType.IDENTITY
 import javax.persistence.Id
 import javax.persistence.Table
-import javax.validation.constraints.NotBlank
-import javax.validation.constraints.NotNull
-import javax.validation.constraints.Past
 
 @Entity
 @Table(name = "teachers")
@@ -32,33 +29,4 @@ class Teacher {
 
 }
 
-data class TeacherResponseDto(
-    val firstName: String, val middleName: String?, val lastName: String, val dateOfBirth: LocalDate
-)
-
 fun Teacher.toTeacherResponseDto() = TeacherResponseDto(firstName!!, middleName, lastName!!, dateOfBirth!!)
-
-data class CreateTeacherDto(
-    @field:NotBlank val firstName: String,
-    val middleName: String?,
-    @field:NotBlank val lastName: String,
-    @field:NotNull @field:Past val dateOfBirth: LocalDate
-)
-
-data class CreateTeacherResponse(val id: Int)
-
-fun Int.toCreateTeacherResponseDto() = CreateTeacherResponse(this)
-
-fun CreateTeacherDto.toTeacher() = Teacher().apply {
-    this.firstName = this@toTeacher.firstName
-    this.middleName = this@toTeacher.middleName
-    this.lastName = this@toTeacher.lastName
-    this.dateOfBirth = this@toTeacher.dateOfBirth
-}
-
-data class PatchTeacherDto(
-    val firstName: String?,
-    val middleName: String?,
-    val lastName: String?,
-    val dateOfBirth: LocalDate?,
-)
