+++ server/src/main/kotlin/com/pearlbailey/pearlbaileyhighschool/teacher/model/TeacherResponseDto.kt
@@ -1,0 +1,7 @@
+package com.pearlbailey.pearlbaileyhighschool.teacher.model
+
+import java.time.LocalDate
+
+data class TeacherResponseDto(
+    val firstName: String, val middleName: String?, val lastName: String, val dateOfBirth: LocalDate
+)
